login complete - frontend

+++ flutter_frontend/projecttaskmanagement/lib/models/projectIntegrate.dart
@@ -6,7 +6,7 @@ class ProjectIntegrate {
     int id;
 
     factory ProjectIntegrate.fromJson(Map<String, dynamic> json) => ProjectIntegrate(
-        id: json["id"],
+        id: json["id"] as int,
     );
 
     Map<String, dynamic> toJson() => {
